+++ nextgisweb/resource/migration/45820e29-add-resource_favorite-table.rw.sql
@@ -1,0 +1,3 @@
+/*** { "revision": "45820e29" } ***/
+
+DROP TABLE resource_favorite;